backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - MySQL
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
alembic==1.13.1

# NLP Libraries
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.4.0
textblob==0.17.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
httpx==0.26.0